+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The following is a curated list of changes in the Enact project, newest changes 
 ### Added
 
 - `core/util` function `checkPropTypes` to check the prop types of a component
+- `spotlight/SpotlightRootDecorator` config `focusEffectClass` to apply a css class to every spottable component when it receives spotlight focus.
+- `spotlight/SpotlightRootDecorator` config `focusEffectClass` and function `setFocusEffectClass` to apply a css class to every spottable component when it receives spotlight focus.
+- `ui/resolution` Added support for linear scaling type
 
 ### Fixed